+++ apm-sniffer/optional-plugins/kotlin-coroutine-plugin/src/main/java/org/apache/skywalking/apm/plugin/kotlin/coroutine/DispatchedTaskRunInterceptor.java
@@ -19,38 +19,55 @@
 package org.apache.skywalking.apm.plugin.kotlin.coroutine;
 
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
+import org.apache.skywalking.apm.agent.core.context.ContextSnapshot;
+import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
+import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 
 import java.lang.reflect.Method;
 
-public class DispatcherInterceptor implements InstanceMethodsAroundInterceptor {
+public class DispatchedTaskRunInterceptor implements InstanceMethodsAroundInterceptor {
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes, MethodInterceptResult result) {
-        if (!ContextManager.isActive()) {
-            return;
-        }
+        if (objInst.getSkyWalkingDynamicField() instanceof ContextSnapshot) {
+            ContextSnapshot snapshot = (ContextSnapshot) objInst.getSkyWalkingDynamicField();
 
-        Runnable runnable = (Runnable) allArguments[1];
+            if (ContextManager.isActive() && snapshot.isFromCurrent()) {
+                // Thread not switched, skip restore snapshot.
+                return;
+            }
 
-        if (Utils.isDispatchedTask(runnable)) {
-            // Using instrumentation for DispatchedContinuation
-            EnhancedInstance enhancedRunnable = (EnhancedInstance) runnable;
-            enhancedRunnable.setSkyWalkingDynamicField(ContextManager.capture());
-        } else {
-            // Wrapping runnable with current context snapshot
-            allArguments[1] = TracingRunnable.wrapOrNot(runnable);
+            // Create local coroutine span
+            AbstractSpan span = ContextManager.createLocalSpan(TracingRunnable.COROUTINE);
+            span.setComponent(ComponentsDefine.KT_COROUTINE);
+            objInst.setSkyWalkingDynamicField(span);
+
+            // Recover with snapshot
+            ContextManager.continued(snapshot);
         }
     }
 
     @Override
     public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes, Object ret) {
+        if (ContextManager.isActive() && objInst.getSkyWalkingDynamicField() instanceof AbstractSpan) {
+            AbstractSpan span = (AbstractSpan) objInst.getSkyWalkingDynamicField();
+            if (span != null) {
+                ContextManager.stopSpan(span);
+            }
+        }
         return ret;
     }
 
     @Override
     public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes, Throwable t) {
+        if (ContextManager.isActive() && objInst.getSkyWalkingDynamicField() instanceof AbstractSpan) {
+            AbstractSpan span = (AbstractSpan) objInst.getSkyWalkingDynamicField();
+            if (span != null) {
+                ContextManager.stopSpan(span.errorOccurred().log(t));
+            }
+        }
     }
 }